add atom data resolver
