+++ tardis/io/util.py
@@ -8,7 +8,11 @@ import pandas as pd
 import numpy as np
 import collections
 from collections import OrderedDict
+
+import requests
 import yaml
+from tqdm._tqdm_notebook import tqdm_notebook
+
 from tardis import constants
 from astropy import units as u
 
@@ -347,5 +351,29 @@ class PlasmaWriterMixin(HDFWriterMixin):
         super(PlasmaWriterMixin, self).to_hdf(file_path, path, name)
 
 
+def download_from_url(url, dst):
+    """
+    kindly used from https://gist.github.com/wy193777/0e2a4932e81afc6aa4c8f7a2984f34e2
+    @param: url to download file
+    @param: dst place to put the file
+    """
 
-
+    file_size = int(requests.head(url).headers["Content-Length"])
+    if os.path.exists(dst):
+        first_byte = os.path.getsize(dst)
+    else:
+        first_byte = 0
+    if first_byte >= file_size:
+        return file_size
+    header = {"Range": "bytes=%s-%s" % (first_byte, file_size)}
+    pbar = tqdm(
+        total=file_size, initial=first_byte,
+        unit='B', unit_scale=True, desc=url.split('/')[-1])
+    req = requests.get(url, headers=header, stream=True)
+    with(open(dst, 'ab')) as f:
+        for chunk in req.iter_content(chunk_size=1024):
+            if chunk:
+                f.write(chunk)
+                pbar.update(1024)
+    pbar.close()
+    return file_size
